Skal bruke state fra søknad når man åpner hovedytelse-siden
