+++ src/frontend/typer/søknad.ts
@@ -1,4 +1,4 @@
-import { AnnenYtelse, Ytelse } from '../barnetilsyn/steg/2-hovedytelse/Hovedytelse';
+import { AnnenYtelse, Ytelse } from '../barnetilsyn/steg/2-hovedytelse/typer';
 
 export interface Hovedytelse {
     ytelse: Exclude<Ytelse, 'annet'> | AnnenYtelse;
